requirements.txt
fastapi
uvicorn
sqlalchemy
google-auth
google-auth-oauthlib
google-api-python-client
openai
